pyproject.toml
[tool.poetry]
name = "public-keys"
version = "0.1.0"
description = ""
authors = ["Andrew Mosson <amosson@yahoo.com>"]
license = "BSD-3-Clause"
packages = [
    { include = "public_keys", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
pynacl = "^1.4.0"
flake8-import-order = "^0.18.1"
flake8-bugbear = "^20.1.4"
saltpack = "^0.2.1"
fusepy = "^3.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
neovim = "^0.3.1"
jedi = "^0.17.0"
jupyter = "^1.0.0"
pynvim = "^0.4.1"
coverage = {extras = ["toml"], version = "^5.1"}
pytest-cov = "^2.9.0"
flake8 = "^3.8.2"
black = "^19.10b0"
flake8-bandit = "^2.1.2"
mypy = "^0.780"
isort = "^4.3.21"
codecov = "^2.1.4"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["public_keys"]

[tool.coverage.report]
show_missing = true
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

# pyproject.toml
